pyproject.toml
[tool.poetry]
name = "Get-Curve-Spot-Price"
version = "0.1.0"
description = "Get the spot price of tokens from by calling the curve.fi swap router"
authors = ["Saikat K saikickkarma@protonmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
eth-ape = "^0.6.26"
structlog = "^23.2.0"
colorama = "^0.4.6"

[tool.poetry.group.lint.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
